create utils file and rename test file
